Change port numnber in mock
We have 2 test files whose test suites have async intensive operations.
It seems I was not able to close the server instances between files
in the the required manner, the end result is that we have one test suite
getting the server instance being binded to port 3000, and then when the other
test file is run it tries to do the same. This result in `PORT ALREADy IN usE`
kind of error. This is by all means a hack, that needs to be fixed.