Remove change simulation in containers/NewPlan
In Planform#L301 the intervention type changeHandler has this statement
setFieldValue('jurisdictions', [initialJurisdictionValues]);
which from the tests undoes the simulated changes that sets
formValues.jurisdiction[0].{id, name}. I removed the statement in the diff
since the interventionType at this point is FI since its the default value.

Might not be very clear.