Downgrade #types/react-redux minor version
Reason: seems the minor versions in v7 do not respect semver rules
There is a breaking change in terms of the exposed interface between
minor versions 7.0.8 and 7.1.7. Specifically the ConnectedComponentClass
type is not exported in 7.1.7. ConnectedComponentClass is used by
onaio/js-tools.

as a result tsd complains that the above type is not present
in node_modules @types/react-redux, which in itself is another
problem since tsd should not be evaluating types in node_modules,
apparently there has been an issue raised for this and its unfortunately
something we cannot fix from here. for Reference:
https://github.com/SamVerschueren/tsd/issues/45