Remove abortFetch functionality from several containers
the goal for abortFetch was to prevent the error of setStating on
an unmounted component. Meaning we only truly need abortFetch in
components whose async operations' callbacks have a setState call.

For most other components the async operations dispatch the response
to the store other than call a setState in the callback. i think
we should retain the callback's functionality since in this case it will
not cause the aforementioned error and also I believe we might want for
resolved data to be added to store if even if we unmount a component.