Feature/772 search (#802)
* Add search functionality for component ActiveFocusInvestigation

Implement search by plan title by filtering case triggered plans and routine plans that match
the search parameter

* Add search for component IRSPlansList

* Refactor search forms into a single reusable component

* Add search for manage plans list

* Debounce search

Delay search to make sure the filter is not applied to frequently

* Add getPlansArrayByTitle selector

* Make use of reselect for Focus investigation search

* Refactor code

* Refactor code to enhance performance

* Refactor code

* Search plan list by title instead of name

* Add selector getPlanDefinitionsArrayByTitle

* Remove plan23 fixture from src/store/ducks/tests/fixtures.ts

* Add selector getIRSPlansArrayByTitle

* Move selectors in ActiveFocusInvestigation to mapStateToProps

Push the searched value as a query param to URL then use that value in selector
to filter by title

* Move search handle submit and handle change to component SearchForm

Have onSubmit push the query string to URL. Remove useDebounce util function
since we are now updating the query string on submit rather than on input
change. Update components to use the enhanced SearchForm

* Add placeholder prop for SearchForm. Update failing test snapshot

* Update SearchForm input value on mount with the value from the query string

If a user searches, the clicks on a record from the returned results, then clicks on the back
browser button, the input is empty but the query string is available in the URL. Update this
value on mount to solve this bug

* Refactor to eliminate redundancy since selectors will handle empty title

* Add default props for component SearchForm

* Remove unused prop for component ActiveFocusInvestigation

* Change query param search to title and access the string from constants

* Add makeIRSPlansArraySelector and use state in mapStateToProps for component IRSPlansList

* Add and use selector makePlanDefinitionsArraySelector

Use selector makePlanDefinitionsArraySelector in place of getPlansArrayByTitle selector. Add tests for selector makeIRSPlansArraySelector

* Use state in place of store.getState() for makePlansArraySelector argument

* Reduce huge PlanDefinitionList snapshot

* Ehance PlanDefinitionList snapshot

* Add HelmetWrapper snapshot for PlanDefinitionList

* Make SearchForm prop placeholder not optional

* Remove repetition