Move WrappedJurisdictionTable back into the PlanAssignment component
This fixes a bug whereby it was not possible to navigate back to a plan's
root jurisdiction because of incorrectly set state within the TreeWalker
component.  With this move, we bypass that by TreeWalker remounting every time.

Sad, but maybe necesary?  Maybe cooler to show a forced loading indicator by default?