Refactor getChildren
We no longer needed the previous incarnation of getChildren.  This is because
we do not have the "limitTree" data structure any more and have switched to using
a jurisdiction tree.  As such, getChildren could be reduced to just an API call to
the findByProperties endpoint.  So I removed it completely.

The current incarnation of getChildren is simply a utility of checking if a node
exists on the tree and if it has children and returning those.