Revert to initialize ReactMapboxGl inside GisidaLite component
The map jankiness addressed by the PR #1021 was fixed by having the Mapbox instance created outside the component to prevent new instances from being created when props change. However, some fill layers stopped being rendered correctly. For instance, some structures that were to be rendered with a fill of yellow are now being rendered with a fill of green. The FI superset data is not available and so a fix to this bug has been flagged as blocked