Add draft for new hierarchies reducer structure
Moves the meta field out of the tree

Meta is saved alongside the tree, we use the jurisdiction Id to know which meta object belongs to  which node in tree.

during selection the final tree is computed(the meta fields are added to thier respective nodes) such that the final tree is similar in structure to how it was previously stored

This means we can update the tree in several service calls without overriding meta information such as to which node has been selected.