Return empty list if row count from superset is 0
If no data is returned from a superset API call, a HTTP status of 400 is returned and superset service returns undefined. This is leads to an error toast been displayed by the caller. In order to work around this, a check  to return an empty list if no rows are returned in the response is implemented in the superset service