Set toDate to time of unassigning team
the opensrp server does not support editing of expired assignments.
so we will use the flow that raises the error; where one assigns a team,
then unassigns the team and tries assigning the team again, all on the same day.
in the second step; if the the toDate is set to an earler date than today, then
if the 3rd step tries doing the same assignment with the same fromDate, the api
will think the user is trying to edit an expired assignment.

Things to remember: there is a combined constraint of plan, organization, jurisdiction, and
fromDate.