Ensure that we can work with jurisdiction metadata where the value is 0
This is because we noticed that the OpenSRP API erroneously returns metadata with no
value.  In this commit we are making it such that our code can deal with these.

Additionally we are using >= to compare with the metadata value instead of just > so that
we can more easily deal with values of 0 and >=100.