Add jurisdictions count
We need to display the jurisdictions count so that the user can get an
indication of how that changes when the slider is adjusted.  This is
especially important in cases where the number of jurisdictions selected
changes but the number of structures selected does not - we want to make
this clear.